Refactor user management components: remove UsersContacts, add UserListUi and ChangeUserFormUi for better user editing and listing functionality in Dashboard and EditTrip pages.
